// crates/lattice-server/frontend/src/generated/ViewLink.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ViewGuestAttachment } from "./ViewGuestAttachment.ts";

export type ViewLink = { id: string, local_device_id: string, local_interface: string, local_ip?: string, remote_device_id: string, remote_interface: string, remote_ip?: string, speed_bps?: number, protocol: string, guest_attachment?: ViewGuestAttachment, network_cidrs: Array<string>, };